Exception handling (some operations) Middleware to convert exceptions

+++ BravaPharm.OrderManagement.Application/Features/Products/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
+using BravaPharm.OrderManagement.Application.Exceptions;
 using BravaPharm.OrderManagement.Application.Interfaces.Persistence;
 using BravaPharm.OrderManagement.Domain.Entities;
 using MediatR;
@@ -23,6 +24,10 @@ namespace BravaPharm.OrderManagement.Application.Features.Products.Commands.Dele
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
             var productToDelete = await _productRepository.GetByIdAsync( request.ProductId );
+            if( productToDelete == null )
+            {
+                throw new NotFoundException(nameof(Product), request.ProductId);
+            }
             await _productRepository.DeleteAsync(productToDelete);
             return Unit.Value;
         }
